==> 01_code/prod/3_MAIN_RUNNER.R <==
#-- This script is run multiple training scripts for different pairs sequentially.
# I left the _CAD and _NZD just for example



rm(list=ls())



source("01_code/prod/3_MAIN_MULTIPLE_CAD.R")
source("01_code/prod/3_MAIN_MULTIPLE_NZD.R")
